PythonUtils/requirements.txt
mypy==1.11.0
mypy-extensions==1.0.0
numpy==2.1.0
parameterized==0.9.0
pillow==10.4.0
pillow_heif==0.18.0
progress==1.6
reverse_geocoder==1.5.1
scipy==1.14.1
typing_extensions==4.12.2
